+++ src/main/java/io/contek/invoker/deribit/api/common/_Ticker.java
@@ -1,16 +1,17 @@
 package io.contek.invoker.deribit.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
+import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Ticker {
 
-  public Double best_ask_amount;
-  public Double best_ask_price;
-  public Double best_bid_amount;
-  public Double best_bid_price;
-  public Double current_funding;
-  public Double funding_8h;
+  public BigDecimal best_ask_amount;
+  public BigDecimal best_ask_price;
+  public BigDecimal best_bid_amount;
+  public BigDecimal best_bid_price;
+  public BigDecimal current_funding;
+  public BigDecimal funding_8h;
   public _Greek greeks;
   public Double index_price;
   public String instrument_name;
